separate reading + writing
f.readlines() will always return [] because you open the file for
writing.

The API claims that the emails you give it should be separated by
commas.

I'm guessing you want to
1. generate the list of emails
2. write them out to the file
3. send them to the API

So I rearranged the work a bit, first generating the list of emails then
writing them with or without newlines and commas as appropriate.

Moved 'template' to the top of the code to separate it from the logic.